Refactor:#76 - standardize plus icon naming, location, and format
- Relocate from public/assets/icons/plusIcon.svg to src/assets/icons/plus.svg
- Apply consistent SVGR props
